refactor(courtManager): 라운드 번호 타입 변경
- roundNumber 타입을 Long에서 Integer로 변경

+++ src/main/java/com/gumraze/drive/drive_backend/courtManager/dto/FreeGameRoundResponse.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Builder
 @Getter
 public class FreeGameRoundResponse {
-    Long roundNumber;
+    Integer roundNumber;
     RoundStatus roundStatus;
     List<FreeGameMatchResponse> matches;
 }
